Split update container and update service into 2 files
To increase readability
