+++ docupdater/update/container.py
@@ -1,70 +1,10 @@
-from abc import ABC, abstractmethod
 from time import sleep
 
 from docker.errors import APIError, NotFound
 
-from .config import Config
+from .abstract import AbstractObject
 from ..helpers.helpers import set_properties, remove_sha_prefix, get_id_from_image
-
-
-class AbstractObject(ABC):
-    def __init__(self, docker_client, object):
-        self.docker = docker_client
-        self.logger = self.docker.logger
-        self.config = self.docker.config
-        self.client = self.docker.client
-        self.socket = self.docker.socket
-
-        self.object = object
-
-    @property
-    def name(self):
-        return self.object.name
-
-    @abstractmethod
-    def get_image_name(self):
-        """Get image name"""
-
-    @abstractmethod
-    def get_tag(self):
-        """Get image tag"""
-
-    @abstractmethod
-    def has_new_version(self):
-        """Return true if there are a new version"""
-
-    @abstractmethod
-    def get_current_id(self):
-        """Return the current id for notification"""
-
-    @abstractmethod
-    def get_latest_id(self):
-        """Return the latest id for notification"""
-
-    def _pull(self, name_with_tag):
-        """Docker pull image tag"""
-        self.logger.debug('Checking tag: %s', name_with_tag)
-        try:
-            if self.config.auth_json:
-                self.client.login(self.config.auth_json.get("username"), self.config.auth_json.get("password"))
-
-            # The authentification doesn't work with this call
-            # See bugs https://github.com/docker/docker-py/issues/2225
-            # return self.client.images.get_registry_data(tag)
-            return self.client.images.pull(name_with_tag, auth_config=self.config.auth_json)
-        except APIError as e:
-            if '<html>' in str(e):
-                self.logger.debug("Docker api issue. Ignoring")
-                raise ConnectionError
-            elif 'unauthorized' in str(e):
-                raise ConnectionError
-            elif 'Client.Timeout' in str(e):
-                self.logger.critical(
-                    "Couldn't find an image on docker.com for %s. Local Build?", name_with_tag)
-                raise ConnectionError
-            elif ('pull access' or 'TLS handshake') in str(e):
-                self.logger.critical("Couldn't pull. Skipping. Error: %s", e)
-                raise ConnectionError
+from ..lib.config import Config
 
 
 class Container(AbstractObject):
@@ -236,83 +176,3 @@ class Container(AbstractObject):
                     self.logger.error('Unable to attach updated container to network "%s". Error: %s', network.name, e)
 
         new_container.start()
-
-
-class Service(AbstractObject):
-    def __init__(self, docker_client, service):
-        super().__init__(docker_client, service)
-        self._latest_sha = None
-        self._current_sha = None
-
-    def get_image_name(self):
-        return self.service.attrs['Spec']['TaskTemplate']['ContainerSpec']['Image'].split(":", 1)[0]
-
-    def get_tag(self):
-        tag = self.service.attrs['Spec']['TaskTemplate']['ContainerSpec']['Image'].split(":", 1)
-        if len(tag) == 2:
-            tag = tag[1]
-            if ":" in tag and "@" in tag:
-                tag = tag.split("@")[0]
-            return tag
-        return "latest"
-
-    def get_sha(self):
-        tag = self.service.attrs['Spec']['TaskTemplate']['ContainerSpec']['Image'].split(":", 1)
-        if len(tag) == 2:
-            tag = tag[1]
-            sha = ""
-            if ":" in tag and "@" in tag:
-                sha = tag.split("@")[1]
-            return remove_sha_prefix(sha)
-        return ""  # Empty, force to update this service for next time
-
-    @property
-    def service(self):
-        return self.object
-
-    def get_current_id(self):
-        return (self._current_sha or "")[:10]
-
-    def get_latest_id(self):
-        return (self._latest_sha or "")[:10]
-
-    def has_new_version(self):
-        self.config = Config.from_labels(self.config, self.service.attrs.get('Spec', dict()).get('Labels'))
-
-        current_image_name = self.get_image_name()
-        current_tag = self.get_tag()
-
-        self.logger.debug("%s:%s", current_image_name, current_tag)
-
-        current_sha = self.get_sha()
-        if not current_sha:
-            self.logger.error('No image SHA for %s. Update it for next time', current_image_name)
-
-        if self.config.latest:
-            current_tag = "latest"
-
-        try:
-            latest_image = self._pull(f"{current_image_name}:{current_tag}")
-        except ConnectionError:
-            return False
-
-        latest_sha = self._get_digest(latest_image)
-        self._latest_sha = latest_sha
-        self._current_sha = current_sha
-        return current_sha != latest_sha
-
-    def update(self):
-        if not self._latest_sha:
-            self.has_new_version()
-
-        self.logger.info('%s will be updated', self.name)
-        self.service.update(image=f"{self.get_image_name()}:{self.get_tag()}@sha256:{self._latest_sha}")
-
-    def _get_digest(self, image):
-        digest = image.attrs.get(
-            "Descriptor", {}
-        ).get("digest") or image.attrs.get(
-            "RepoDigests"
-        )[0].split('@')[1] or image.id
-
-        return remove_sha_prefix(digest)
